consumer/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
boto3==1.34.40
python-dotenv==1.0.1
pandas==2.2.0
